requirements.txt
Django==4.2
gunicorn==20.0.4
djangorestframework==3.15.1
psycopg2==2.9.9

sqlparse==0.3.1
Pillow
